Added Automatic BNLS Server Finder. It currently uses the address http://stealthbot.net/p/bnls.txt, so it won't work until a file is uploaded there.
It ignores the first line, and then treats each line from then on as a BNLS server. I left the first line vacant so after v2.7, we could possibly use it to improve connection errors people get by placing and using verbyte details or the such there.

I also added an option in the Advanced Settings tab to disable this feature.