Changed Database Manager UI:   Removed the 'What is This?' Button   Moved the Tabs to something more pleasing looking
Changed CreateOBj() function to check if the Object name is  valid VB name. Meaning Only A-Z0-9_ and 1st Char A-Z. 
CreateObj() now has the possibility to return Nothing, so be prepared.

Implemented Reason grabbing from Banned messages, For clsBanlistUserObj

Whisper Window can now be Enabled/Disabled when the window is Minimized/Maximized

Fixed infinite loop in DBManager when adding a new user when a "New_User_#1" already existed in the list.

Watching user names can now have Wildcards.

Exposed The User database to scripting in the function GetUserDatabase() Retruns a collection Who's Items are clsDBEntryObjs and who's Keys are the names

Modified objDBEntrys to have support for Groups, and Ban Messages

Add function to Scripting, GetCommands() returns a collection who's Items are clsCommandDocsObjs, and Keys are the command names

When loading the User database, it will now automatically convert old 2.6r3 style User.txts to the new format.