Complete overhaul of config system
Created clsConfig, a class dedicated to handling the config file. It's
responsible for reading, writing, and storing values. All of the
previous config handlign code has been removed (except for some things
which are planned to be taken out of the config).

Along with this comes an entirely new config format, version 6. This
version is much more organized. A backup of any previous version config
file will be made automatically prior to upgrading the format.

Given that the new system caches changes rather than writing them
immediately, the entire file must be saved in order to write any
changes. When saving, only modified values are written, and performance
during these saves seems to be affected minimally. These saves occur
after any config change (menu options and /setX commands).

This system is still in ALPHA and requires a lot more testing.

Also with this update I've removed a bunch of unused or legacy code.