Outgoing filters no longer affect internal commands
They are not used on any messages beginning with / unless it is a
whisper, ban, kick, or away message. In these cases, the filter is not
applied to the command itself or the target user (if there is one), only
the actual messages or kick/ban reason.

Also removed a bunch of old input processing code that's no longer being
used.