Updated settings form panels (adv. conn.)
- Removed outdated information regarding BNLS
- Removed check box for the BNLS finder. The option can now be found in
the server list, as the default and only "server" provided with a fresh
install. The finder will only be enabled if this option is selected,
selecting any other option will disable it.
- Moved the check box for "Connect on Startup" from the misc settings
panel to the advanced connection settings panel.
- Added a text box to set a default email address used with new account
registrations.
- Added a tooltip to the hash file path label (useful for long,
uninterrupted paths). I also made it so that when a path is longer than
a certain number of characters (80, 100) it decreases the font size.
- Changed the text for ping spoofing options to be less redundant.

Additionally, some defaults were changed in the config, the saving of
the BNLS server list file was tweaked, old code was removed, and
existing code was re-arranged and commented on.

The bot will no longer attempt to connect to BNLS if a server is not
set. If the server finder is enabled, it will try and find one first,
otherwise it will give an error message.

Tab indexes have changed on the settings form, but they are still in
order on each panel.