Standardized product info
Created a UDT "ProductInfo" and built an array of these types which
contains the basic information about every product (4-digit code, key,
number of CD keys, full name, BNLS ID, logon system, and default version
byte). Updated all functions using this information to get it from this
array. There is also an entry ProductList(0) which is the "unknown"
product.

The global "GetProductInfo(str)" function can be used to get a product's
information. str can be either the 4-character product code, it's
reverse, or the short code.