Fixed BNLS server box behavior
If the user manually inputs a server, it is now correctly saved and
added to the list.

If the finder is disabled and no server is set, a "No server set" value
will be displayed.

The server selection box is now disabled when hashing is selected as the
connection method.