Changed 'phrases' command
Added an optional argument for getting the number of banned phrases. If
no argument is specified, the command will list all of the phrases,
unless more than 3 messages would be output, in which case it will only
output the number of phrases. The sub-command 'list' can be used to
forcibly list all of the phrases.