Reset window position when number of monitors changes
Two new config items under UI-Position. EnforceBounds is a bool which
determines if we should check for an out of bounds window. MonitorCount
is the number of monitors on the computer the last time the bot was ran.
If the number of monitors changes from the last time, the window will
reset to the primary monitor.