Updated old installer
The NSIS (?) thing that was used for the 488 release wasn't working out.

This version is much larger than previous ones due to the inclusion of
the Visual C runtimes (found in Installer\Modules), which are required
for Warden.dll. It also contains updated component libraries.

This installer can be used to "upgrade" older installations (I only
tested 486, but if the upgrade code was the same it should be fine for
any other version).

Also removed the other installer's files.