Add FullName property to user object
Returns the name of the user and their gateway. Follows the same
conversion rules as using the "Show All" namespace convention.

Also added optional arguments to modCommands.ConvertUsername and
modCommands.ConvertUsernameGateway to allow a convention to be specified
other than the bot's current one. This number is the same as the
conversion type used in these functions.
0 = none, 1 = legacy, 2 = wc3, 3 = all