Improve SID_READUSERDATA handling
User data requests are now added to an array which links the requested
keys and account name to the returned values. This replaces what was
described by comments as "some of the oldest code in the project".

A new scripting event was added to include all of this information,
Event_UserDataReceived(account, keys(), values()). A new function,
RequestUserData(account, keys()) is also added to the SSC for requesting
multiple keys at once. Any request will trigger both the new event and
the old KeyReturn event.

This also fixes #23.

Internally, modParsing.RequestSpecificKey() has been replaced by
RequestUserData, which allows a type and command to be specified. There
are 4 types, Internal, ProfileWindow, ScriptingCall, and UserCommand. A
response to a ProfileWindow request will open the profile window,
UserCommand will attempt to reply to the optional command object,
Internal will be displayed in the bot window. ScriptingCall will only
trigger the events.

These changes should also properly handle the filetime strings returned
by the server for certain system keys.