Require intent when deleting group through command
If you use the /rem command and the affected entry would be a group, you
now have to explicitly specify the entry type (using --type group). The
user will be prompted if they do not do this.

Perhaps this could be improved to seek out users first before groups but
this will do for now.

Finishes implementing #37
Part of the database overhaul included adding support for the --type
argument to the remove command.