Fix optional timestamp passing in logger
IsMissing() only works on Variants. This changes the TimeDate parameter
for all of the WriteX functions on the logger class to variant and
implements the check in each function. If the parameter is ever ommitted
at any stage, the current time will be used.

The change was also made to the modPacketBuffer.MakePacket() function,
though the optional parameter is never supplied in practice. This fixes
#64