Improve void-view performance
- Changed refresh time to 60 seconds (either after joining if no users
received or after the last user is processed).
- Re-arranged flag update event handler so that the channel list isn't
iterated over needlessly.
- The VoidView setting now only controls whether automatic /unignore
commands are sent and if users discovered by flag updates are added to
the visible channel list. Event_FlagUpdate is raised even if this is
disabled and the internal user list (Channel.Users()) is updated
normally. This doesn't appear to have any noticeable performance impact
given the new packet processing method.