More consistent FreeFile() calls.
It was missing in someplaces. I did a check for 'open x for y as #z'
calls and now it's called before each open, only when a file is actually
going to be opened.

There was also one case in the SSC where static file numbers were used
that now uses FreeFile().

This should fix  some of the problems from issue #59.