Tweak phantom user handling
If a user in channel event is received for a user already present, with
no changes to the stats, the original user is treated as a phantom. If a
user joins with the same name as someone already in the channel, the
original is treated as a phantom. When a phantom is discovered, they are
removed from the channel entirely and the new user replaces them. If
this happens, an alert will be printed once per channel visit.

Phantom users seem to be mostly gone on the server so this is completely
untested.